+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- The Dashboard's six-month spending chart now shows the current month at the top instead of the bottom.
+
 ## [0.4.1] - 2026-07-23
 
 ### Changed